% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{check_pkg}
\alias{check_pkg}
\title{Check basic installed packages}
\usage{
check_pkg(pkgs)
}
\arguments{
\item{pkgs}{charachter vector of package(s) to check for presence}

\item{...}{further options for install_fun}
}
\value{
invisible character vector with missing package
}
\description{
Check basic installed packages
}
