% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_to_xml.R
\name{cost_to_xml}
\alias{cost_to_xml}
\title{Convert cost data to XML}
\usage{
cost_to_xml(cost_data, file = NULL)
}
\arguments{
\item{cost_data}{a data frame of cost data, i.e. individual cost per
event considered.}

\item{file}{[chr] (default is NULL) if provided as
\code{"path/to/xml/output/xml"} the function write the
XML in this file (note: folder must be exists).}
}
\value{
an (invisible) XML document
}
\description{
This function aims to convert cost data from data a frmae to XML
format.
}
\examples{
\dontrun{
  library(imthcm)
  data(meps_costs)
  xml_cost <- 'test_cost.xml'
  cost_to_xml(meps_costs, file = xml_cost)
  converted_costs <- xml_to_cost(file = xml_cost)
  identical(converted_costs, meps_costs)
}
}
