% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_models_ita}
\alias{event_models_ita}
\title{Default italiantrained models.}
\format{A list of 3 list each of lenght 7:
\describe{
  \item{summer}{[GENERAL DESCRIPTION FOR SUMMER]
    \describe{
      \item{mort_all}{...}
      \item{mort_cardiac}{...}
      \item{mort_resp}{...}
      \item{mort_cer}{...}
      \item{hosp_cardiac}{...}
      \item{hosp_resp}{...}
      \item{hosp_cer}{...}
    }
  }
  \item{non_summer}{[GENERAL DESCRIPTION FOR SUMMER]
    \describe{
      \item{mort_all}{...}
      \item{mort_cardiac}{...}
      \item{mort_resp}{...}
      \item{mort_cer}{...}
      \item{hosp_cardiac}{...}
      \item{hosp_resp}{...}
      \item{hosp_cer}{...}
    }
  }
  \item{full_year}{[GENERAL DESCRIPTION FOR SUMMER]
    \describe{
      \item{mort_all}{...}
      \item{mort_cardiac}{...}
      \item{mort_resp}{...}
      \item{mort_cer}{...}
      \item{hosp_cardiac}{...}
      \item{hosp_resp}{...}
      \item{hosp_cer}{...}
    }
  }
}}
\usage{
event_models_ita
}
\description{
[GENERAL DESCRIPTION FOR THE MODELS, ONECE DONE DELETE ALSO THE SQUARE BRACKETS]
}
\keyword{datasets}
