% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_me.R
\name{parse_version}
\alias{parse_version}
\title{Parse version of a package}
\usage{
parse_version(ver)
}
\arguments{
\item{ver}{character string representing the version of a package}
}
\value{
an integer vector with the component of the version
}
\description{
Parse version of a package
}
\examples{
imthcm:::parse_version('0.2.1')
}
