% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_computed_cost.R
\name{plot_computed_costs}
\alias{plot_computed_costs}
\title{Plot number of costs associated to predicted number of health events}
\usage{
plot_computed_costs(computed_costs, plot_file)
}
\arguments{
\item{computed_costs}{[data frame] A data frame containing the
following variables:
     - `date`: date of simulated day in format 'yyyy-mm-dd';
     - `event`: name of the outcomes;
     - `lower_individual_cost`: lower bound of 95% CI of costs
        associated to a particular health event;
     - `fit_individual_cost`: costs associated to the average daily
        number of a particular health event;
     - `upper_individual_cost`: upper bound of 95% CI of costs
        associated to a particular health event.}

\item{plot_file}{[chr] filename of the destination path. Format of image
is automatically decided by the filename extension}
}
\value{
A boxplot showing the costs associated to the predicted number
        of a particular health event.
}
\description{
A plot that shows the number of average daily costs associated to
a particular health event. The bold line represents the
point estimate of the mean value of associated to a particular health
event. The crossbar represents the 95% confidence interval.
}
\examples{
\dontrun{
  library(imthcm)
  default_models <- train_event_models(use_ita = TRUE)

  hh_non_summer <- predict_hm(models = default_models,
    weather_history = test_weather,
    weather_today   = test_weather[731L, ]
  )

  hh_full_year <- predict_hm(models = default_models,
    weather_history = test_weather,
    weather_today   = test_weather[c(731L, 730L), ],
    full_year = TRUE
  )

  computed_costs <- compute_cost(hh_non_summer, use_meps = TRUE)
  plot_computed_costs(computed_costs,
    plot_file = 'costs_plot.png'
  )
}
}
