% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred_event_outcomes_time.R
\name{plot_pred_event_outcomes_time}
\alias{plot_pred_event_outcomes_time}
\title{Plot number of predicted events in function of time}
\usage{
plot_pred_event_outcomes_time(plot_data, plot_file)
}
\arguments{
\item{plot_data}{[data frame] A data frame containing the
following variables:
     - `date` : date of simulated day in format 'yyyy-mm-dd';
     - `event`: name of the outcomes;
     - `lower`: lower bound of 95% CI of average daily predicted number
                of events;
     - `fit`  : average daily predicted number of events;
     - `upper`: upper bound of 95% CI of average daily predicted number
                of events.}

\item{plot_file}{[chr] filename of the destination path. Format of image
is automatically decided by the filename extension}
}
\value{
A plot with time trends of the average daily number of predicted
        health outcomes considered.
}
\description{
A plot that shows the number of predicted events in function of time.
The points represent the predicted average daily number of health events.
The colours refer to the type of outcome.
}
\examples{
\dontrun{
  library(imthcm)
  default_models <- train_event_models(use_ita = TRUE)
  predicted_events <- predict_hm(default_models,
    test_weather, test_weather[-c(1, 2, 3), ]
  )
  plot_pred_event_outcomes_time(predicted_events,
    plot_file = 'timeoutcome_plot.png'
  )
}
}
