% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_events.R
\name{predict_hm}
\alias{predict_hm}
\title{Predict function for Health module}
\usage{
predict_hm(models, weather_history, weather_today, ..., full_year = (!"o38h"
  \%in\% names(weather_today)) || (all(purrr::map_lgl(models[["summer"]],
  is.null))))
}
\arguments{
\item{models}{[lst] TO BE COMMENTED}

\item{weather_history}{[data frame] TO BE COMMENTED}

\item{weather_today}{[data frame] TO BE COMMENTED}

\item{...}{possible further arguments passed to the function}

\item{full_year}{[lgl] should be models (and prediction) made on the
same full-year data or model should distinguish from summer
period and non-summer period? If \code{TRUE} full-year models
are used. By default, full-year are used in the case of missing
summer- or non-summer-model or if \code{weather_history} or
\code{weather_today} do not have ozone information (i.e., there
are no \code{o38h} variable)}
}
\value{
a data frame with the fitted value (i.e., mean event predicted)
        including 95% Confidence Interval (column \code{lower} and
        \code{upper}) for each type of event mange by the models used for
        each date considered.
}
\description{
Predict function for Health module
}
\examples{
library(imthcm)
default_models <- train_event_models(use_ita = TRUE)
predict_hm(default_models, test_weather, test_weather[731L, ])
predict_hm(default_models, test_weather, test_weather[c(731L, 730L), ],
  full_year = TRUE
)

no_o3_test_weather <- dplyr::select(test_weather, -o38h)
predict_hm(default_models, no_o3_test_weather,
  no_o3_test_weather[731L, ]
)
}
