% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_me.R
\name{remote_version}
\alias{remote_version}
\title{Get the version of the package in the remote}
\usage{
remote_version(con)
}
\arguments{
\item{con}{a connection to the remote DESCRIPTION file of the package}
}
\value{
a character string with the version of the package
}
\description{
Get the version of the package in the remote
}
\examples{

url(
  'https://raw.githubusercontent.com/UBESP-DCTV/imthcm/master/DESCRIPTION'
) \%>\%
imthcm:::remote_version()
}
