% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_health}
\alias{test_health}
\title{Test dataset for health input.}
\format{A data frame with 1827 rows and 8 variables:
\describe{
  \item{date}{...}
  \item{mort_all}{...}
  \item{mort_cardiac}{...}
  \item{mort_resp}{...}
  \item{mort_cer}{...}
  \item{hosp_cardiac}{...}
  \item{hosp_resp}{...}
  \item{hosp_cer}{...}
}}
\usage{
test_health
}
\description{
A dataset containing ...
}
\keyword{datasets}
