% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_weather}
\alias{test_weather}
\title{Test dataset for weather input.}
\format{A data frame with 1827 rows and 7 variables:
\describe{
  \item{date}{...}
  \item{temp_mean}{...}
  \item{press_bar_mean}{...}
  \item{pm10}{...}
  \item{pm25}{...}
  \item{no2}{...}
  \item{o38h}{...}
}}
\usage{
test_weather
}
\description{
A dataset containing ...
}
\keyword{datasets}
