% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_event_models.R
\name{train_event_models}
\alias{train_event_models}
\title{Events model trainer}
\usage{
train_event_models(health_events_history = NULL, weather_history = NULL,
  ..., use_ita = FALSE)
}
\arguments{
\item{health_events_history}{A data frame with historical health events
data for at least one week each month of at least one year (full
years are obviusly preferrend) are needed to take into
account the stagionality. Colum names must be included in:
- mort_all     = for daily number of death (any causes);
- mort_cardiac = for daily number of death (cardiac causes only);
- mort_resp    = for daily number of death
  (respiratory casues only);
- mort_cer     = for daily number of death
  (cerebrovascular causes only);
- hosp_cardiac = for daily number of hospitalization (cardiac
  causes only);
- hosp_resp    = for daily number of hospitalization
  (respiratory casues only);
- hosp_cer     = for daily number of hospitalization
  (cerebrovascular causes only);}

\item{weather_history}{[data frame] A data frame with weather historical
       data with number of rows equal to the lenght of
       \code{health_events_history} and at least the following column
       (with exactly the same column names):
       - date           = date expressed in the format 'yyyy-mm-dd';
       - temp_mean      = mean temperature of corresponding day
                          (Celsius);
       - press_bar_mean = mean pressure of corresponding day (hPascal);
       - pm10           = mean value of pm10 of corresponding day
                          (\eqn{\mu g/m^3});
       - pm25           = mean value of pm25 of corresponding day
                          (\eqn{\mu g/m^3});
       - no2            = mean value of no2  of corresponding day
                          (\eqn{\mu g/m^3}).

       If provided, by a column named \code{038h}, also information
       of O3 daily maximum concentration for the daily 8-hours
       moving-mean will be used (only for summer period, i.e., from
       April, 1st, to September, 30th).}

\item{...}{Other possible options passed to the function}

\item{use_ita}{[lgl] (default = FALSE) use italian historical
data on weather if the user cannot provide more specific data.}
}
\value{
a list of three sets of gam models, one set with the summer model
        for each type ov event provided in \code{health_events_history}
        (taking into account also information from O3), one set for the
        non-summer models (using only non-O3 pollutant information) and
        a set of full-year models which do not use information on O3
        (if not provided). Together all the model of the first two sets
        or the models of the third set are alternative \code{NULL}.
        Output is returned \code{invisible()}. See Details section for
        more informations.
}
\description{
A function that trains models to estimates the average daily number of
events.
}
\details{
The output is a list of three lists, named \code{summer},
\code{non_summer} and \code{full_year}. The first two are usefull if
ozone data are provided to assess a more specific model considering
also ozone in summer period and only other pollutants in the rest of the
year. The last set (i.e., list) of models are trained and aimed to be
used without any knowledge about ozone concentrations.
}
\examples{
\dontrun{
  library(imthcm)
  test_models    <- train_event_models(test_health, test_weather)
  default_models <- train_event_models(use_ita = TRUE)
}
}
