% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_me.R
\name{update_me}
\alias{update_me}
\title{Update imthcm package itself}
\usage{
update_me()
}
\value{
invisible()
}
\description{
Aim of this function is to check and eventually update the imthcm package
with a possible new version on GitHub.
}
\examples{
library(imthcm)
update_me()
}
