% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_preproc.R
\name{weather_preproc}
\alias{weather_preproc}
\title{Pre-process function}
\usage{
weather_preproc(weather_data)
}
\arguments{
\item{weather_data}{[data frame] A data.frame}
}
\value{
A data.frame
}
\description{
A function that pre-process data to be take as inputs by GAM models.
}
\examples{
library(imthcm)
weather_preproc(test_weather)
}
