% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_to_xml.R
\name{weather_to_xml}
\alias{weather_to_xml}
\title{Convert weather data to XML}
\usage{
weather_to_xml(weather_data, file = NULL)
}
\arguments{
\item{weather_data}{a data frame of historical weather data.}

\item{file}{[chr] (default is NULL) if provided as
\code{"path/to/xml/output/xml"} the function write the
XML in this file (note: folder must be exists).}
}
\value{
an (invisible) XML document
}
\description{
This function aims to convert weather data from data a frmae to XML
format.
}
\examples{
library(imthcm)
\dontrun{
  data(test_weather)
  xml_weather <- 'test_weather.xml'
  weather_to_xml(test_weather, file = xml_weather)
  weather_history <- xml_to_weather(file = xml_weather)
  identical(weather_history, test_weather)
}
}
