% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{west_data}
\alias{west_data}
\title{West street data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 624 rows and 17 columns.}
\usage{
west_data
}
\description{
Data from the Padova's West street
}
\keyword{datasets}
