% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_cost.R
\name{xml_to_cost}
\alias{xml_to_cost}
\title{Read an XML cost data}
\usage{
xml_to_cost(file)
}
\arguments{
\item{file}{[chr] path to the XML cost file to import}
}
\value{
a data frame of cost data
}
\description{
Read an XML cost data
}
\examples{
\dontrun{
  library(imthcm)
  data(meps_costs)
  xml_cost <- 'test_cost.xml'
  cost_to_xml(meps_costs, file = xml_cost)
  converted_costs <- xml_to_cost(file = xml_cost)
  identical(converted_costs, meps_costs)
}
}
