% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_health.R
\name{xml_to_health}
\alias{xml_to_health}
\title{Read an XML Health data}
\usage{
xml_to_health(file)
}
\arguments{
\item{file}{[chr] path to the XML health file to import}
}
\value{
a data frame of health data
}
\description{
Read an XML Health data
}
\examples{
library(imthcm)
\dontrun{
  data(test_health)
  xml_health <- 'test_health.xml'
  health_to_xml(test_health, file = xml_health)
  events_history <- xml_to_health(file = xml_health)
  identical(events_history, test_health)
}
}
