% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_prediction.R
\name{xml_to_prediction}
\alias{xml_to_prediction}
\title{Read an XML predicted data}
\usage{
xml_to_prediction(file)
}
\arguments{
\item{file}{[chr] path to the XML predicted file to import (e.g. the
output of hm.R)}
}
\value{
a data frame of predicted data
}
\description{
Read an XML predicted data
}
\examples{
\dontrun{
  library(imthcm)
  default_models <- train_event_models(use_ita = TRUE)
  pred <- predict_hm(default_models,
    test_weather, test_weather[c(730L, 731L), ]
  )
  predictions_to_xml(pred, file = 'predictions.xml')
  imported_pred <- xml_to_prediction(file = 'predictions.xml')
  identical(pred, imported_pred)
}
}
