% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_weather.R
\name{xml_to_weather}
\alias{xml_to_weather}
\title{Read an XML weather data}
\usage{
xml_to_weather(file)
}
\arguments{
\item{file}{[chr] path to the XML weather file to import}
}
\value{
a data frame of weather data
}
\description{
Read an XML weather data
}
\examples{
library(imthcm)
\dontrun{
  data(test_weather)
  xml_weather <- 'test_weather.xml'
  weather_to_xml(test_weather, file = xml_weather)
  weather_history <- xml_to_weather(file = xml_weather)
  identical(weather_history, test_weather)
}
}
