% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_to_xml.R
\name{cost_to_xml}
\alias{cost_to_xml}
\title{Convert cost data to XML}
\usage{
cost_to_xml(cost_data, file = NULL)
}
\arguments{
\item{cost_data}{a data frame of cost data, i.e. individual cost per
event considered.}

\item{file}{[chr] (default is NULL) if provided as
\code{"path/to/xml/output/xml"} the function write the
XML in this file (note: folder must be exists).}
}
\value{
an (invisible) XML document
}
\description{
This function aims to convert cost data from data a frmae to XML
format.
}
\examples{
\dontrun{
  library(imthcm)
  data(test_health)
  xml_health <- 'test_health.xml'
  health_to_xml(test_health, file = xml_health)
  events_history <- xml_to_health(file = xml_health)
  identical(events_history, test_health)
}
}
