% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_health}
\alias{test_health}
\title{Test dataset for health input.}
\format{A data frame with 1827 rows and 8 variables:
\describe{
  \item{date}{date of the day}
  \item{mort_all}{number of deaths for all causes}
  \item{mort_cardiac}{number of deaths for cardiac diseases}
  \item{mort_resp}{number of deaths for respiratory diseases}
  \item{mort_cer}{number of deaths for cerebrovascular diseases}
  \item{hosp_cardiac}{number of hospitalizations for cardiac diseases}
  \item{hosp_resp}{number of hospitalizations for respiratory diseases}
  \item{hosp_cer}{
    number of hospitalizations for cerebrovascular diseases
  }
}}
\usage{
test_health
}
\description{
A dataset containing simulated historical information on daily average
number of health events. Number of deaths for all causes, cardiac
diseases, respiratory diseases and cerebrovascular diseases are the
mortality events considered. Number of hospitalizations for cardiac
diseases, respiratory diseases and cerebrovascular diseases are the
hospitalization events considered.
}
\keyword{datasets}
