% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proc.R
\name{weather_preproc}
\alias{weather_preproc}
\title{Pre-process function}
\usage{
weather_preproc(weather_data)
}
\arguments{
\item{weather_data}{[data frame] A data.frame}
}
\value{
A data.frame
}
\description{
descrizione breve (a function that pre-process data to be take as inputs
by GAM models.)
}
\details{
Se serve (dal secondo paragrafo in poi) ci sono i dettagli
}
\examples{
imthcm:::weather_preproc(test_weather)
}
